requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.3.2
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cryptography==3.4.8
docutils==0.17.1
filelock==3.6.0
Flask==3.0.0
Flask-WTF==1.2.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gunicorn==20.1.0
httplib2==0.20.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
protobuf==4.25.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.1
urllib3==1.26.5
Werkzeug==3.0.1
WTForms==3.1.1
